Extract preflight into its own service package.
Move orchestration out of package service so later checker work has a home without changing HTTP, JSON, or activation behavior.
