+++ internal/service/deconfliction/deconfliction_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/Aero-Arc/aero-arc-api/internal/domain"
 	. "github.com/Aero-Arc/aero-arc-api/internal/service"
 	. "github.com/Aero-Arc/aero-arc-api/internal/service/deconfliction"
+	. "github.com/Aero-Arc/aero-arc-api/internal/service/preflight"
 	"github.com/Aero-Arc/aero-arc-api/internal/store/durable"
 	durablememory "github.com/Aero-Arc/aero-arc-api/internal/store/durable/memory"
 )
